add CreateCompletionReportAsync CalculateReportCostAsync modify models CompletionReport & CompletionReportOperation

+++ GrainElevatorAPI/DTO/Requests/UpdateRequests/CompletionReportOperationUpdateRequest.cs
@@ -7,6 +7,6 @@ public class CompletionReportOperationUpdateRequest
     [Range(1, int.MaxValue, ErrorMessage = "TechnologicalOperationId must be a positive number.")]
     public int? TechnologicalOperationId { get; set; }
 
-    [Range(0, double.MaxValue, ErrorMessage = "Amount must be a positive number.")]
-    public double? Amount { get; set; }
+    [Range(0, int.MaxValue, ErrorMessage = "Amount must be a positive number.")]
+    public int? Amount { get; set; }
 }
